requirements.txt
crypto==1.4.1
pycryptodome==3.20.0
pillow==10.4.0
